refactor: vite.config.js 의 jsxInject 옵션을 통해 전역으로 함수 등록하는 로직 제거
